requirements.txt
pyspark==3.5.0
pandas>=2.0.0
PyYAML>=6.0
pytest>=7.0.0
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0


